refactor(334): allow import type for ui
