--- features/profile/ui/components/UserDataBox/UserDataBox.logic.ts
@@ -1,1 +1,0 @@
-export type { Trip } from '@/features/trips';
