+++ features/profile/ui/components/UserDataBox/UserDataBox.tsx
@@ -1,5 +1,5 @@
-import type { Trip } from '@/features/profile/ui/components/UserDataBox/UserDataBox.logic';
 import { styles } from '@/features/profile/ui/components/UserDataBox/UserDataBox.style';
+import type { Trip } from '@/features/trips/domain/entities/Trip';
 import { CustomText } from '@/ui/components/basic/CustomText/CustomText';
 import { type FC, Fragment } from 'react';
 import { Pressable, View } from 'react-native';
