--- features/trip-generation/ui/components/SummaryCard/SummaryCard.logic.ts
@@ -1,1 +1,0 @@
-export type { TravelerInfo } from '@/features/trip-generation/domain/entities/TravelerInfo';
