--- features/trips/ui/components/DayItem/DayItem.logic.ts
@@ -1,1 +1,0 @@
-export type { DayPlan } from '@/features/trips/domain/entities/DayPlan';
