+++ features/trips/ui/components/DayItem/DayItem.tsx
@@ -1,5 +1,5 @@
+import type { DayPlan } from '@/features/trips/domain/entities/DayPlan';
 import { ActivityItem } from '@/features/trips/ui/components/ActivityItem/ActivityItem';
-import type { DayPlan } from '@/features/trips/ui/components/DayItem/DayItem.logic';
 import { styles } from '@/features/trips/ui/components/DayItem/DayItem.style';
 import type { FC } from 'react';
 import { FlatList, View } from 'react-native';
