--- features/trips/ui/components/FoodCard/FoodCard.logic.ts
@@ -1,1 +1,0 @@
-export type { Food } from '@/features/trips/domain/entities/Food';
