+++ features/trips/ui/components/FoodCard/FoodCard.tsx
@@ -1,4 +1,4 @@
-import type { Food } from '@/features/trips/ui/components/FoodCard/FoodCard.logic';
+import type { Food } from '@/features/trips/domain/entities/Food';
 import { styles } from '@/features/trips/ui/components/FoodCard/FoodCard.style';
 import { CustomIcon } from '@/ui/components/basic/CustomIcon/CustomIcon';
 import { CustomText } from '@/ui/components/basic/CustomText/CustomText';
