--- features/trips/ui/components/ListHeaderComponent/ListHeaderComponent.logic.ts
@@ -1,3 +1,0 @@
-export type { Food } from '@/features/trips/domain/entities/Food';
-export type { TripDetails } from '@/features/trips/domain/entities/TripDetails';
-export type { Weather } from '@/features/trips/domain/entities/Weather';
