+++ features/trips/ui/components/ListHeaderComponent/ListHeaderComponent.tsx
@@ -1,9 +1,7 @@
+import type { Food } from '@/features/trips/domain/entities/Food';
+import type { TripDetails } from '@/features/trips/domain/entities/TripDetails';
+import type { Weather } from '@/features/trips/domain/entities/Weather';
 import { FoodCard } from '@/features/trips/ui/components/FoodCard/FoodCard';
-import type {
-  Food,
-  TripDetails,
-  Weather,
-} from '@/features/trips/ui/components/ListHeaderComponent/ListHeaderComponent.logic';
 import { style } from '@/features/trips/ui/components/ListHeaderComponent/ListHeaderComponent.style';
 import { MapListHeaderComponent } from '@/features/trips/ui/components/MapListHeaderComponent/MapListHeaderComponent';
 import { NotesCard } from '@/features/trips/ui/components/NotesCard/NotesCard';
