--- features/trips/ui/components/WeatherCard/WeatherCard.logic.ts
@@ -1,1 +1,0 @@
-export type { Weather } from '@/features/trips/domain/entities/Weather';
