+++ features/trips/ui/pages/TripDetailsPage/TripDetailsPage.tsx
@@ -1,8 +1,8 @@
 import { Routes } from '@/features/core/navigation';
+import type { DayPlan } from '@/features/trips/domain/entities/DayPlan';
 import { DayItem } from '@/features/trips/ui/components/DayItem/DayItem';
 import { HeaderIcons } from '@/features/trips/ui/components/HeaderIcons/HeaderIcons';
 import { ListHeaderComponent } from '@/features/trips/ui/components/ListHeaderComponent/ListHeaderComponent';
-import type { DayPlan } from '@/features/trips/ui/pages/TripDetailsPage/TripDetailsPage.logic';
 import { useTripDetailsPageLogic } from '@/features/trips/ui/pages/TripDetailsPage/TripDetailsPage.logic';
 import { styles } from '@/features/trips/ui/pages/TripDetailsPage/TripDetailsPage.style';
 import { AnimatedHeaderImage } from '@/ui/components/composite/AnimatedHeaderImage/AnimatedHeaderImage';
