+++ features/trips/ui/pages/TripListPage/TripListPage.logic.ts
@@ -3,9 +3,6 @@ import { useGetTrips } from '@/features/trips/facades/useGetTrips';
 import { useUniqueItems } from '@/features/trips/hooks/useUniqueItems';
 import { useEffect } from 'react';
 
-export type { Trip } from '@/features/trips/domain/entities/Trip';
-export type { UniqueItem } from '@/features/trips/hooks/useUniqueItems';
-
 export const useTripListPageLogic = () => {
   const { getUniqueItems } = useUniqueItems();
   const { isLoading, totalTrips, trips } = useGetTrips();
